Moved base class to its own file and refactored a variable name
